Updated tictactoe.py
I added the basic 2 player functionality and it works. Now, all that is left is changing some of the print messages so that it corresponds to the game being 2 player. Hopefully, you can try doing this. Run the program, evaluate what you think needs to be done, and try making changes. I think we will be finishing this project within the next couple of days.